Backend/requirements.txt
fastapi>=0.115,<0.116
uvicorn>=0.34,<0.35
langchain>=0.3.25,<0.4
chromadb>=1.0,<2.0
openai>=1.78,<2.0
pydantic>=2.11,<3
pypdf>=3.1               # en lugar de PyPDF2
tqdm>=4.67
python-dotenv>=1.1,<2
